requirements.txt
azure-functions
requests==2.32.3
vaderSentiment==3.3.2
